+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.16] - 2026-03-28
+
+### Docs
+- Update examples/complex_demo/README.md
+
+### Other
+- Update examples/complex_demo/data/users.csv
+- Update examples/complex_demo/main.py
+
 ## [0.1.15] - 2026-03-28
 
 ### Other